requirements.txt
streamlit
openai
nltk
spacy
pypdf
textstat
textblob
python-dotenv
llama-index==0.10.1
llama-index-agent-openai==0.1.1
llama-index-core==0.10.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.1
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.1